Fix: Replace "#help" to reflect #3
In #3 I added the ability to use a custom prefix, since # is a poor prefix to use for a bot, and changing that required editing StashDumper.py. This commit updates the documentation to reflect this change. While "{prefix}help" isn't the most elegant way, it could cause any confusion made by a user typing #help when their prefix is different.